+++ CHANGELOG.md
@@ -29,6 +29,8 @@ Copyright (c) 2022-2026 Claudio Satriano <satriano@ipgp.fr>
   time-formatted axis/colorbar when using `time`
 - Added `--colorby` support to map plots, as well as `--colormap` support
   (with autocompletion) for `seiscat plot` and `seiscat timeline`
+- In `seiscat plot` the option `-m/--maptype` has been renamed to
+  `-b/--backend` for consistency with `seiscat timeline`
 - `seiscat download` renamed to `seiscat fetchdata`
 - New option for `seiscat fetchdata`: `--sds` to retrieve waveform data from
   a local SDS archive